requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-telegram-bot==21.9
openai==1.59.6
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
